+++ ParallelSystems/meanshift-cuda/include/serial.h
@@ -7,10 +7,6 @@
 #include <float.h>
 #include <math.h>
 
-#include "../include/global_vars.h"
-#include "../include/serial.h"
-#include "../include/helpers.h"
-
 void serial(void);
 
 void init_serial(void);
@@ -21,7 +17,7 @@ void free_memory(void);
 
 void read_file(void);
 
-void write_csv_file (char *message, long double **a, const int ROW, const int COL);
+void write_csv_file (char *message, double **a, const int ROW, const int COL);
 
 int validate(void);
 
@@ -35,18 +31,18 @@ void matrix_mult(void);
 
 void normalize(void);
 
-long double sum_of_row(const int row_index);
+double sum_of_row(const int row_index);
 
-long double frob_norm(void);
+double frob_norm(void);
 
 void calc_meanshift(void);
 
-void copy_2Darray(long double **source, long double **destination, const int ROW, const int COL);
+void copy_2Darray(double **source, double **destination, const int ROW, const int COL);
 
-void print_2Darray(long double **a, const int ROW, const int COL);
+void print_2Darray(double **a, const int ROW, const int COL);
 
-long double gaussian_kernel(const long double dist);
+double gaussian_kernel(const double dist);
 
-long double euclidean_distance(const int first, const int second);
+double euclidean_distance(const int first, const int second);
 
 #endif
